Move BuildCommon into its own file
Defaults.scala was very slow to compile and format due to the sheer
amount of code contained in it. We can reduce the compile time by
splitting it into multiple files. This should be a no-op functionaly.
