+++ main/src/main/scala/sbt/Defaults.scala
@@ -8,7 +8,7 @@
 package sbt
 
 import java.io.{ File, PrintWriter }
-import java.net.{ URI, URL }
+import java.net.URL
 import java.nio.file.{ Paths, Path => NioPath }
 import java.util.Optional
 import java.util.concurrent.TimeUnit
@@ -16,7 +16,7 @@ import java.util.concurrent.TimeUnit
 import lmcoursier.CoursierDependencyResolution
 import org.apache.logging.log4j.core.{ Appender => XAppender }
 import org.scalasbt.ipcsocket.Win32SecurityLevel
-import sbt.Def.{ Initialize, ScopedKey, Setting, SettingsDefinition }
+import sbt.Def.{ Initialize, ScopedKey, Setting }
 import sbt.Keys._
 import sbt.Project.{
   inConfig,
@@ -26,7 +26,7 @@ import sbt.Project.{
   richTaskSessionVar,
   sbtRichTaskPromise,
 }
-import sbt.Scope.{ GlobalScope, ThisScope, fillTaskAxis }
+import sbt.Scope.{ GlobalScope, ThisScope }
 import sbt.coursierint._
 import sbt.internal.CommandStrings.ExportStream
 import sbt.internal._
@@ -70,7 +70,6 @@ import sbt.testing.{ AnnotatedFingerprint, Framework, Runner, SubclassFingerprin
 import sbt.util.CacheImplicits._
 import sbt.util.InterfaceUtil.{ toJavaFunction => f1, t2 }
 import sbt.util._
-import sjsonnew._
 
 import scala.collection.immutable.ListMap
 import scala.concurrent.duration._
@@ -2285,343 +2284,7 @@ object Defaults extends BuildCommon {
     }
 }
 
-private[sbt] object Build0 extends BuildExtra
-
-trait BuildExtra extends BuildCommon with DefExtra {
-  import Defaults._
-
-  /**
-   * Defines an alias given by `name` that expands to `value`.
-   * This alias is defined globally after projects are loaded.
-   * The alias is undefined when projects are unloaded.
-   * Names are restricted to be either alphanumeric or completely symbolic.
-   * As an exception, '-' and '_' are allowed within an alphanumeric name.
-   */
-  def addCommandAlias(name: String, value: String): Seq[Setting[State => State]] = {
-    val add = (s: State) => BasicCommands.addAlias(s, name, value)
-    val remove = (s: State) => BasicCommands.removeAlias(s, name)
-    def compose(setting: SettingKey[State => State], f: State => State) =
-      setting in GlobalScope ~= (_ compose f)
-    Seq(compose(onLoad, add), compose(onUnload, remove))
-  }
-
-  /**
-   * Adds Maven resolver plugin.
-   */
-  def addMavenResolverPlugin: Setting[Seq[ModuleID]] =
-    libraryDependencies += sbtPluginExtra(
-      ModuleID("org.scala-sbt", "sbt-maven-resolver", sbtVersion.value),
-      sbtBinaryVersion.value,
-      scalaBinaryVersion.value
-    )
-
-  /**
-   * Adds `dependency` as an sbt plugin for the specific sbt version `sbtVersion` and Scala version `scalaVersion`.
-   * Typically, use the default values for these versions instead of specifying them explicitly.
-   */
-  def addSbtPlugin(
-      dependency: ModuleID,
-      sbtVersion: String,
-      scalaVersion: String
-  ): Setting[Seq[ModuleID]] =
-    libraryDependencies += sbtPluginExtra(dependency, sbtVersion, scalaVersion)
-
-  /**
-   * Adds `dependency` as an sbt plugin for the specific sbt version `sbtVersion`.
-   * Typically, use the default value for this version instead of specifying it explicitly.
-   */
-  def addSbtPlugin(dependency: ModuleID, sbtVersion: String): Setting[Seq[ModuleID]] =
-    libraryDependencies += {
-      val scalaV = (scalaBinaryVersion in update).value
-      sbtPluginExtra(dependency, sbtVersion, scalaV)
-    }
-
-  /**
-   * Adds `dependency` as an sbt plugin for the sbt and Scala versions configured by
-   * `sbtBinaryVersion` and `scalaBinaryVersion` scoped to `update`.
-   */
-  def addSbtPlugin(dependency: ModuleID): Setting[Seq[ModuleID]] =
-    libraryDependencies += {
-      val sbtV = (sbtBinaryVersion in pluginCrossBuild).value
-      val scalaV = (scalaBinaryVersion in update).value
-      sbtPluginExtra(dependency, sbtV, scalaV)
-    }
-
-  /** Transforms `dependency` to be in the auto-compiler plugin configuration. */
-  def compilerPlugin(dependency: ModuleID): ModuleID =
-    dependency.withConfigurations(Some("plugin->default(compile)"))
-
-  /** Adds `dependency` to `libraryDependencies` in the auto-compiler plugin configuration. */
-  def addCompilerPlugin(dependency: ModuleID): Setting[Seq[ModuleID]] =
-    libraryDependencies += compilerPlugin(dependency)
-
-  /** Constructs a setting that declares a new artifact `a` that is generated by `taskDef`. */
-  def addArtifact(a: Artifact, taskDef: TaskKey[File]): SettingsDefinition = {
-    val pkgd = packagedArtifacts := packagedArtifacts.value updated (a, taskDef.value)
-    Seq(artifacts += a, pkgd)
-  }
-
-  /** Constructs a setting that declares a new artifact `artifact` that is generated by `taskDef`. */
-  def addArtifact(
-      artifact: Initialize[Artifact],
-      taskDef: Initialize[Task[File]]
-  ): SettingsDefinition = {
-    val artLocal = SettingKey.local[Artifact]
-    val taskLocal = TaskKey.local[File]
-    val art = artifacts := artLocal.value +: artifacts.value
-    val pkgd = packagedArtifacts := packagedArtifacts.value updated (artLocal.value, taskLocal.value)
-    Seq(artLocal := artifact.value, taskLocal := taskDef.value, art, pkgd)
-  }
-
-  def externalIvySettings(
-      file: Initialize[File] = inBase("ivysettings.xml"),
-      addMultiResolver: Boolean = true
-  ): Setting[Task[IvyConfiguration]] =
-    externalIvySettingsURI(file(_.toURI), addMultiResolver)
-
-  def externalIvySettingsURL(
-      url: URL,
-      addMultiResolver: Boolean = true
-  ): Setting[Task[IvyConfiguration]] =
-    externalIvySettingsURI(Def.value(url.toURI), addMultiResolver)
-
-  def externalIvySettingsURI(
-      uri: Initialize[URI],
-      addMultiResolver: Boolean = true
-  ): Setting[Task[IvyConfiguration]] = {
-    val other = Def.task {
-      (
-        baseDirectory.value,
-        appConfiguration.value,
-        projectResolver.value,
-        updateOptions.value,
-        streams.value
-      )
-    }
-    ivyConfiguration := ((uri zipWith other) {
-      case (u, otherTask) =>
-        otherTask map {
-          case (base, app, pr, uo, s) =>
-            val extraResolvers = if (addMultiResolver) Vector(pr) else Vector.empty
-            ExternalIvyConfiguration()
-              .withLock(lock(app))
-              .withBaseDirectory(base)
-              .withLog(s.log)
-              .withUpdateOptions(uo)
-              .withUri(u)
-              .withExtraResolvers(extraResolvers)
-        }
-    }).value
-  }
-
-  private[this] def inBase(name: String): Initialize[File] = Def.setting {
-    baseDirectory.value / name
-  }
-
-  def externalIvyFile(
-      file: Initialize[File] = inBase("ivy.xml"),
-      iScala: Initialize[Option[ScalaModuleInfo]] = scalaModuleInfo
-  ): Setting[Task[ModuleSettings]] =
-    moduleSettings := IvyFileConfiguration(
-      ivyValidate.value,
-      iScala.value,
-      file.value,
-      managedScalaInstance.value
-    )
-
-  def externalPom(
-      file: Initialize[File] = inBase("pom.xml"),
-      iScala: Initialize[Option[ScalaModuleInfo]] = scalaModuleInfo,
-  ): Setting[Task[ModuleSettings]] =
-    moduleSettings := PomConfiguration(
-      ivyValidate.value,
-      iScala.value,
-      file.value,
-      managedScalaInstance.value,
-    )
-
-  def runInputTask(
-      config: Configuration,
-      mainClass: String,
-      baseArguments: String*
-  ): Initialize[InputTask[Unit]] =
-    Def.inputTask {
-      import Def._
-      val r = (runner in (config, run)).value
-      val cp = (fullClasspath in config).value
-      val args = spaceDelimited().parsed
-      r.run(mainClass, data(cp), baseArguments ++ args, streams.value.log).get
-    }
-
-  def runTask(
-      config: Configuration,
-      mainClass: String,
-      arguments: String*
-  ): Initialize[Task[Unit]] =
-    Def.task {
-      val cp = (fullClasspath in config).value
-      val r = (runner in (config, run)).value
-      val s = streams.value
-      r.run(mainClass, data(cp), arguments, s.log).get
-    }
-
-  // public API
-  /** Returns a vector of settings that create custom run input task. */
-  def fullRunInputTask(
-      scoped: InputKey[Unit],
-      config: Configuration,
-      mainClass: String,
-      baseArguments: String*
-  ): Vector[Setting[_]] = {
-    // TODO: Re-write to avoid InputTask.apply which is deprecated
-    // I tried "Def.spaceDelimited().parsed" (after importing Def.parserToInput)
-    // but it broke actions/run-task
-    // Maybe it needs to be defined inside a Def.inputTask?
-    @com.github.ghik.silencer.silent
-    def inputTask[T](f: TaskKey[Seq[String]] => Initialize[Task[T]]): Initialize[InputTask[T]] =
-      InputTask.apply(Def.value((s: State) => Def.spaceDelimited()))(f)
-
-    Vector(
-      scoped := inputTask { result =>
-        initScoped(
-          scoped.scopedKey,
-          ClassLoaders.runner mapReferenced Project.mapScope(s => s.in(config))
-        ).zipWith(Def.task { ((fullClasspath in config).value, streams.value, result.value) }) {
-          (rTask, t) =>
-            (t, rTask) map {
-              case ((cp, s, args), r) =>
-                r.run(mainClass, data(cp), baseArguments ++ args, s.log).get
-            }
-        }
-      }.evaluated
-    ) ++ inTask(scoped)(forkOptions in config := forkOptionsTask.value)
-  }
-
-  // public API
-  /** Returns a vector of settings that create custom run task. */
-  def fullRunTask(
-      scoped: TaskKey[Unit],
-      config: Configuration,
-      mainClass: String,
-      arguments: String*
-  ): Vector[Setting[_]] =
-    Vector(
-      scoped := initScoped(
-        scoped.scopedKey,
-        ClassLoaders.runner mapReferenced Project.mapScope(s => s.in(config))
-      ).zipWith(Def.task { ((fullClasspath in config).value, streams.value) }) {
-          case (rTask, t) =>
-            (t, rTask) map {
-              case ((cp, s), r) =>
-                r.run(mainClass, data(cp), arguments, s.log).get
-            }
-        }
-        .value
-    ) ++ inTask(scoped)(forkOptions in config := forkOptionsTask.value)
-
-  def initScoped[T](sk: ScopedKey[_], i: Initialize[T]): Initialize[T] =
-    initScope(fillTaskAxis(sk.scope, sk.key), i)
-  def initScope[T](s: Scope, i: Initialize[T]): Initialize[T] =
-    i mapReferenced Project.mapScope(Scope.replaceThis(s))
-
-  /**
-   * Disables post-compilation hook for determining tests for tab-completion (such as for 'test-only').
-   * This is useful for reducing test:compile time when not running test.
-   */
-  def noTestCompletion(config: Configuration = Test): Setting[_] =
-    inConfig(config)(Seq(definedTests := detectTests.value)).head
-
-  def filterKeys(ss: Seq[Setting[_]], transitive: Boolean = false)(
-      f: ScopedKey[_] => Boolean
-  ): Seq[Setting[_]] =
-    ss filter (s => f(s.key) && (!transitive || s.dependencies.forall(f)))
-}
-
 trait DefExtra {
   private[this] val ts: TaskSequential = new TaskSequential {}
   implicit def toTaskSequential(@deprecated("unused", "") d: Def.type): TaskSequential = ts
-}
-
-trait BuildCommon {
-
-  /**
-   * Allows a String to be used where a `NameFilter` is expected.
-   * Asterisks (`*`) in the string are interpreted as wildcards.
-   * All other characters must match exactly.  See GlobFilter.
-   */
-  implicit def globFilter(expression: String): NameFilter = GlobFilter(expression)
-
-  implicit def richAttributed(s: Seq[Attributed[File]]): RichAttributed = new RichAttributed(s)
-  implicit def richFiles(s: Seq[File]): RichFiles = new RichFiles(s)
-  implicit def richPathFinder(s: PathFinder): RichPathFinder = new RichPathFinder(s)
-  final class RichPathFinder private[sbt] (s: PathFinder) {
-
-    /** Converts the `PathFinder` to a `Classpath`, which is an alias for `Seq[Attributed[File]]`. */
-    def classpath: Classpath = Attributed blankSeq s.get
-  }
-  final class RichAttributed private[sbt] (s: Seq[Attributed[File]]) {
-
-    /** Extracts the plain `Seq[File]` from a Classpath (which is a `Seq[Attributed[File]]`).*/
-    def files: Seq[File] = Attributed.data(s)
-  }
-  final class RichFiles private[sbt] (s: Seq[File]) {
-
-    /** Converts the `Seq[File]` to a Classpath, which is an alias for `Seq[Attributed[File]]`. */
-    def classpath: Classpath = Attributed blankSeq s
-  }
-
-  def overrideConfigs(cs: Configuration*)(
-      configurations: Seq[Configuration]
-  ): Seq[Configuration] = {
-    val existingName = configurations.map(_.name).toSet
-    val newByName = cs.map(c => (c.name, c)).toMap
-    val overridden = configurations map { conf =>
-      newByName.getOrElse(conf.name, conf)
-    }
-    val newConfigs = cs filter { c =>
-      !existingName(c.name)
-    }
-    overridden ++ newConfigs
-  }
-
-  // these are intended for use in in put tasks for creating parsers
-  def getFromContext[T](task: TaskKey[T], context: ScopedKey[_], s: State): Option[T] =
-    SessionVar.get(SessionVar.resolveContext(task.scopedKey, context.scope, s), s)
-
-  def loadFromContext[T](task: TaskKey[T], context: ScopedKey[_], s: State)(
-      implicit f: JsonFormat[T]
-  ): Option[T] =
-    SessionVar.load(SessionVar.resolveContext(task.scopedKey, context.scope, s), s)
-
-  // intended for use in constructing InputTasks
-  def loadForParser[P, T](task: TaskKey[T])(
-      f: (State, Option[T]) => Parser[P]
-  )(implicit format: JsonFormat[T]): Initialize[State => Parser[P]] =
-    loadForParserI(task)(Def value f)(format)
-  def loadForParserI[P, T](task: TaskKey[T])(
-      init: Initialize[(State, Option[T]) => Parser[P]]
-  )(implicit format: JsonFormat[T]): Initialize[State => Parser[P]] =
-    Def.setting { (s: State) =>
-      init.value(s, loadFromContext(task, resolvedScoped.value, s)(format))
-    }
-
-  def getForParser[P, T](
-      task: TaskKey[T]
-  )(init: (State, Option[T]) => Parser[P]): Initialize[State => Parser[P]] =
-    getForParserI(task)(Def value init)
-  def getForParserI[P, T](
-      task: TaskKey[T]
-  )(init: Initialize[(State, Option[T]) => Parser[P]]): Initialize[State => Parser[P]] =
-    Def.setting { (s: State) =>
-      init.value(s, getFromContext(task, resolvedScoped.value, s))
-    }
-
-  // these are for use for constructing Tasks
-  def loadPrevious[T](task: TaskKey[T])(implicit f: JsonFormat[T]): Initialize[Task[Option[T]]] =
-    Def.task { loadFromContext(task, resolvedScoped.value, state.value)(f) }
-  def getPrevious[T](task: TaskKey[T]): Initialize[Task[Option[T]]] =
-    Def.task { getFromContext(task, resolvedScoped.value, state.value) }
-
-  private[sbt] def derive[T](s: Setting[T]): Setting[T] =
-    Def.derive(s, allowDynamic = true, trigger = _ != streams.key, default = true)
 }
